+++ src/features/url-shortener/functions.ts
@@ -1,6 +1,6 @@
-interface ShortenUrlResponse {
-    url: string;
-}
+import { db } from '@/app/api/client';
+
+import { ShortenUrlResponse } from './type';
 
 export const shortenUrl = async (url: string) => {
     const response = await fetch('/api/url-entry', {
@@ -18,3 +18,15 @@ export const shortenUrl = async (url: string) => {
     const data = (await response.json()) as ShortenUrlResponse;
     return data;
 };
+
+export async function getOriginalUrl(slug: string) {
+    const urlEntry = await db.urlEntry.findUnique({
+        where: { slug: slug },
+    });
+
+    if (urlEntry === null) {
+        throw new Error('URL not found');
+    }
+
+    return urlEntry.originalUrl;
+}
