+++ src/features/url-shortener/type.ts
@@ -1,0 +1,3 @@
+export interface ShortenUrlResponse {
+    url: string;
+}
